backend/requirements.txt
annotated-types==0.5.0
anyio==3.7.1
bcrypt==4.0.1
click==8.1.7
dnspython==2.4.2
email-validator==2.0.0.post2
exceptiongroup==1.1.3
fastapi==0.103.1
h11==0.14.0
idna==3.4
MarkupSafe==2.1.3
motor==3.3.1
passlib==1.7.4
pydantic==2.3.0
pydantic_core==2.6.3
PyJWT==2.8.0
pymongo==4.5.0
python-decouple==3.8
python-dotenv==1.0.0
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.7.1
uvicorn==0.23.2
